edit

+++ src/main/java/kg/attractor/job_search_java_25/repository/EducationInfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EducationInfoRepository extends JpaRepository<EducationInfo, Integer> {
+public interface EducationInfoRepository extends JpaRepository<EducationInfo, Long> {
 
 }
